+++ tests/MessageTest.php
@@ -7,7 +7,9 @@ namespace Yiisoft\Log\Tests;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\InvalidArgumentException;
 use Psr\Log\LogLevel;
+use stdClass;
 use Yiisoft\Log\Message;
+use Yiisoft\Log\Tests\TestAsset\StringableObject;
 
 final class MessageTest extends TestCase
 {
@@ -52,29 +54,113 @@ final class MessageTest extends TestCase
         new Message('unknown', 'message');
     }
 
-    public function parseMessageProvider(): array
+    public function dataParseMessage(): array
     {
         return [
-            [
+            'no-placeholder' => [
                 'no placeholder',
                 ['foo' => 'some'],
                 'no placeholder',
             ],
-            [
+            'string' => [
                 'has {foo} placeholder',
                 ['foo' => 'some'],
                 'has some placeholder',
             ],
-            [
+            'without-value' => [
                 'has {foo} placeholder',
                 [],
                 'has {foo} placeholder',
+            ],
+            'null' => [
+                'has "{foo}" placeholder',
+                ['foo' => null],
+                'has "" placeholder',
+            ],
+            'array' => [
+                'has "{foo}" placeholder',
+                ['foo' => ['bar' => 7]],
+                <<<TEXT
+                has "[
+                    'bar' => 7
+                ]" placeholder
+                TEXT,
+            ],
+            'nested' => [
+                'has "{foo.bar}" placeholder',
+                ['foo' => ['bar' => 7]],
+                'has "7" placeholder',
+            ],
+            'nested-non-exist' => [
+                'has "{foo.bar}" placeholder',
+                ['foo' => []],
+                'has "{foo.bar}" placeholder',
+            ],
+            'nested-non-stringable' => [
+                'has "{foo.bar}" placeholder',
+                ['foo' => new stdClass()],
+                'has "{foo.bar}" placeholder',
+            ],
+            'stringable' => [
+                'has "{foo}" placeholder',
+                ['foo' => new StringableObject('test')],
+                'has "test" placeholder',
+            ],
+            'nested-placeholder' => [
+                'has "{a{b}c}" placeholder',
+                ['a{b}c' => 'test'],
+                'has "test" placeholder',
+            ],
+            'nested-quoted' => [
+                'has "{foo\.ba\\\\r}" placeholder',
+                ['foo.ba\\r' => 'test'],
+                'has "test" placeholder',
+            ],
+            'nested-extended-1' => [
+                'has "{foo\\\.bar}" placeholder',
+                ['foo\\' => ['bar' => 'test']],
+                'has "test" placeholder',
+            ],
+            'nested-extended-2' => [
+                'has "{foo\\\\\\\\.bar}" placeholder',
+                ['foo\\\\' => ['bar' => 'test']],
+                'has "test" placeholder',
+            ],
+            'nested-extended-3' => [
+                'has "{foo\\\.}" placeholder',
+                ['foo\\' => ['' => 'test']],
+                'has "test" placeholder',
+            ],
+            'nested-extended-4' => [
+                'has "{foo\\\\}" placeholder',
+                ['foo\\' => 'test'],
+                'has "test" placeholder',
+            ],
+            'nested-extended-5' => [
+                'has "{foo\.bar.a}" placeholder',
+                ['foo.bar' => ['a' => 'test']],
+                'has "test" placeholder',
+            ],
+            'nested-extended-6' => [
+                'has "{key1\..\.key2\..\.key3}" placeholder',
+                ['key1.' => ['.key2.' => ['.key3' => 'test']]],
+                'has "test" placeholder',
+            ],
+            'nested-extended-7' => [
+                'has "{key1\..\.key2\..\.key3}" placeholder',
+                ['key1.' => ['.key2.' => ['.key3' => 'test']]],
+                'has "test" placeholder',
+            ],
+            'empty' => [
+                'Value — "{}"',
+                ['' => 'test'],
+                'Value — "test"',
             ],
         ];
     }
 
     /**
-     * @dataProvider parseMessageProvider
+     * @dataProvider dataParseMessage
      */
     public function testParseMessage(string $message, array $context, string $expected): void
     {